+++ srcs/structures/engine/player/getters_setters/get_pos_x.c
@@ -6,13 +6,13 @@
 /*   By: ncoudsi <ncoudsi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/07/14 13:16:35 by ncoudsi           #+#    #+#             */
-/*   Updated: 2020/07/14 13:16:53 by ncoudsi          ###   ########.fr       */
+/*   Updated: 2020/09/02 09:57:31 by ncoudsi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cube.h"
 
-int	pos_x()
+float	pos_x()
 {
 	return (g_engine->player->pos->x);
 }
